+++ govapp/frontend/pbs/src/components/forms/colocation/datatable_template.vue
